requirements.txt
discord.py ~= 1.5.1
pyyaml ~= 5.3.1
GitPython ~= 3.1.11
asyncpg ~= 0.21.0
